server/requirements.txt
asgiref==3.3.4
Django==3.2
djangorestframework==3.12.4
psycopg2-binary==2.8.6
pytz==2021.1
Pillow==8.0.0
sqlparse==0.4.1
django-filter==2.4.0
